Move new pipeline to main src dir
